This branch introduces parallel algorithms in Triangulation_3 and Mesh_3,
as well as a Concurrent_compact_container. The corresponding features are:
https://cgal.geometryfactory.com/CGAL/Members/wiki/Features/Parallel_triangulation_3
https://cgal.geometryfactory.com/CGAL/Members/wiki/Features/Parallel_Mesh_3

Tested in CGAL-4.5-Ic-98.
Approved by the Release Manager.

+++ Triangulation_3/benchmark/Triangulation_3/simple.cpp
@@ -28,7 +28,7 @@ int main()
   }
   Timer timer;
   timer.start();
-  int N = 0;
+  size_t N = 0;
   for(int i = 0; i < 5; i++){
     DT dt;
     dt.insert(points.begin(), points.end());
